zimage/pyproject.toml
[project]
name = "zimage-service"
version = "0.1.0"
description = "Zenoh-based Python service for AI inference with FlatBuffers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "zenoh",
    "flatbuffers>=23.0.0",
    "diffusers>=0.26.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "pillow>=10.0.0",
    "accelerate>=0.25.0",
    "huggingface-hub>=0.19.0"
]
